# perf: Faster bitpacking for Parquet writer