--- a/crates/polars-parquet/Cargo.toml
+++ b/crates/polars-parquet/Cargo.toml
@@ -25,7 +25,6 @@ polars-utils = { workspace = true }
 simdutf8 = { workspace = true }
 
 parquet-format-safe = "0.2"
-rand = "0.8"
 seq-macro = { version = "0.3", default-features = false }
 streaming-decompression = "0.1"
 
@@ -40,6 +39,9 @@ snap = { version = "^1.1", optional = true }
 zstd = { version = "^0.13", optional = true, default-features = false }
 
 xxhash-rust = { version = "0.8", optional = true, features = ["xxh64"] }
+
+[dev-dependencies]
+rand = "0.8"
 
 [features]
 compression = [