--- a/crates/polars-parquet/src/lib.rs
+++ b/crates/polars-parquet/src/lib.rs
@@ -1,4 +1,6 @@
+#![feature(test)]
 #![allow(clippy::len_without_is_empty)]
 pub mod arrow;
 pub use crate::arrow::{read, write};
 pub mod parquet;
+extern crate test;