--- a/crates/polars-parquet/src/parquet/encoding/bitpacked/encode.rs
+++ b/crates/polars-parquet/src/parquet/encoding/bitpacked/encode.rs
@@ -1,4 +1,4 @@
-use super::{Packed, Unpackable, Unpacked};
+use super::{Unpackable, Unpacked};
 
 /// Encodes (packs) a slice of [`Unpackable`] into bitpacked bytes `packed`, using `num_bits` per value.
 ///
@@ -42,7 +42,7 @@ pub fn encode<T: Unpackable>(unpacked: &[T], num_bits: usize, packed: &mut [u8])
 /// Only the first `ceil8(unpacked.len() * num_bits)` of `packed` are populated.
 #[inline]
 pub fn encode_pack<T: Unpackable>(unpacked: &[T], num_bits: usize, packed: &mut [u8]) {
-    if unpacked.len() < T::Packed::LENGTH {
+    if unpacked.len() < T::Unpacked::LENGTH {
         let mut complete_unpacked = T::Unpacked::zero();
         complete_unpacked.as_mut()[..unpacked.len()].copy_from_slice(unpacked);
         T::pack(&complete_unpacked, num_bits, packed)