I replaced the Parquet bit-packing algorithm with a modified version of the scalar algorithm from https://github.com/quickwit-oss/bitpacking (unsure where/how to give credit). I ensured the new algorithm works by adding unit tests that encode/decode random data.

I added a microbenchmark in my first commit to help measure the performance gain.

|| Before  | After |
|-| ------------- | ------------- |
|Avg. (ns/iter)|93.75 | 4.25 |
|Std. (ns/iter)| 1.57 | 0.04 |

I removed the microbenchmark in a later commit because it requires nightly Rust and is probably not useful outside of this PR.

Other miscellaneous fixes:
- Packed array of 64 u64's should have a maximum length of 64 * 8 u8's (second commit)
- Input array length should be compared to expected unpacked length instead of expected packed length to avoid unnecessary allocations/copies (third commit)