@ritchie46 2024-05-18
Nice improvement! Thanks a lot @thalassemia. :raised_hands:  Hope you find some on the reading side as well. ;)